+++ query_control_test.go
@@ -209,7 +209,7 @@ func (s *RethinkSuite) TestControlError(c *test.C) {
 	c.Assert(err, test.NotNil)
 	c.Assert(err, test.FitsTypeOf, RQLUserError{})
 
-	c.Assert(err.Error(), test.Equals, "gorethink: An error occurred in: \nr.Error(\"An error occurred\")")
+	c.Assert(err.Error(), test.Equals, "gorethink: An error occurred in:\nr.Error(\"An error occurred\")")
 }
 
 func (s *RethinkSuite) TestControlDoNothing(c *test.C) {