+++ query_table.go
@@ -14,7 +14,7 @@ type TableCreateOpts struct {
 	NonVotingReplicaTags interface{} `gorethink:"nonvoting_replica_tags,omitempty"`
 }
 
-func (o *TableCreateOpts) toMap() map[string]interface{} {
+func (o TableCreateOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }
 
@@ -68,7 +68,7 @@ type IndexCreateOpts struct {
 	Geo   interface{} `gorethink:"geo,omitempty"`
 }
 
-func (o *IndexCreateOpts) toMap() map[string]interface{} {
+func (o IndexCreateOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }
 
@@ -123,7 +123,7 @@ type IndexRenameOpts struct {
 	Overwrite interface{} `gorethink:"overwrite,omitempty"`
 }
 
-func (o *IndexRenameOpts) toMap() map[string]interface{} {
+func (o IndexRenameOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }
 
@@ -159,7 +159,7 @@ type ChangesOpts struct {
 }
 
 // ChangesOpts contains the optional arguments for the Changes term
-func (o *ChangesOpts) toMap() map[string]interface{} {
+func (o ChangesOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }