+++ query_transformation.go
@@ -13,10 +13,10 @@ import p "gopkg.in/dancannon/gorethink.v2/ql2"
 //     })
 func Map(args ...interface{}) Term {
 	if len(args) > 0 {
-		args = append(args[:len(args)-1], funcWrapArgs(args[len(args)-1:])...)
+		args = append(args[:len(args)-1], funcWrap(args[len(args)-1]))
 	}
 
-	return constructRootTerm("Map", p.Term_MAP, funcWrapArgs(args), map[string]interface{}{})
+	return constructRootTerm("Map", p.Term_MAP, args, map[string]interface{}{})
 }
 
 // Map transforms each element of the sequence by applying the given mapping
@@ -28,7 +28,11 @@ func Map(args ...interface{}) Term {
 //         return row.Mul(2)
 //     })
 func (t Term) Map(args ...interface{}) Term {
-	return constructMethodTerm(t, "Map", p.Term_MAP, funcWrapArgs(args), map[string]interface{}{})
+	if len(args) > 0 {
+		args = append(args[:len(args)-1], funcWrap(args[len(args)-1]))
+	}
+
+	return constructMethodTerm(t, "Map", p.Term_MAP, args, map[string]interface{}{})
 }
 
 // WithFields takes a sequence of objects and a list of fields. If any objects in the
@@ -52,7 +56,7 @@ type OrderByOpts struct {
 	Index interface{} `gorethink:"index,omitempty"`
 }
 
-func (o *OrderByOpts) toMap() map[string]interface{} {
+func (o OrderByOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }
 
@@ -111,7 +115,7 @@ type SliceOpts struct {
 	RightBound interface{} `gorethink:"right_bound,omitempty"`
 }
 
-func (o *SliceOpts) toMap() map[string]interface{} {
+func (o SliceOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }
 
@@ -156,7 +160,7 @@ type UnionOpts struct {
 	Interleave interface{} `gorethink:"interleave,omitempty"`
 }
 
-func (o *UnionOpts) toMap() map[string]interface{} {
+func (o UnionOpts) toMap() map[string]interface{} {
 	return optArgsToMap(o)
 }